requirements.txt
flask==3.0.3
tweepy==4.14.0
praw==7.7.1
vaderSentiment==3.3.2
textblob==0.18.0.post0
requests==2.32.3
ccxt==4.3.67
scikit-learn==1.5.1
pandas==2.2.2
numpy==2.0.1
matplotlib==3.9.0
cryptography==43.0.0  # For key encryption
anthropic==0.28.0  # For portfolio optimization
python-dotenv==1.0.0
